Remove stdlib argument to backend initialization, separate stdlib required by compiled reach programs
