+++ examples/tut-1/index.mjs
@@ -13,11 +13,11 @@ import * as backend from './build/index.main.mjs';
 
   await Promise.all([
     backend.Alice(
-      stdlib, ctcAlice,
+      ctcAlice,
       {},
     ),
     backend.Bob(
-      stdlib, ctcBob,
+      ctcBob,
       {},
     ),
   ]);
