+++ examples/tut-2/index.mjs
@@ -26,11 +26,11 @@ import * as backend from './build/index.main.mjs';
 
   await Promise.all([
     backend.Alice(
-      stdlib, ctcAlice,
+      ctcAlice,
       Player('Alice'),
     ),
     backend.Bob(
-      stdlib, ctcBob,
+      ctcBob,
       Player('Bob'),
     ),
   ]);
