+++ examples/tut-4-attack/index.mjs
@@ -29,11 +29,11 @@ import * as backend from './build/index.main.mjs';
   });
 
   await Promise.all([
-    backend.Alice(stdlib, ctcAlice, {
+    backend.Alice(ctcAlice, {
       ...Player('Alice'),
       wager: stdlib.parseCurrency(5),
     }),
-    backend.Bob(stdlib, ctcBob, {
+    backend.Bob(ctcBob, {
       ...Player('Bob'),
       acceptWager: (amt) => {
         console.log(`Bob accepts the wager of ${fmt(amt)}.`);
