+++ js/ts/ALGO.ts
@@ -1,3 +1,7 @@
+// ****************************************************************************
+// standard library for Javascript users
+// ****************************************************************************
+
 // XXX: do not import any types from algosdk; instead copy/paste them below
 // XXX: can stop doing this workaround once @types/algosdk is shippable
 import algosdk from 'algosdk';
@@ -10,31 +14,26 @@ import {
   CurrencyAmount, OnProgress, WPArgs,
   debug, getDEBUG,
   isBigNumber, bigNumberify,
-  mkAddressEq, makeDigest, argsSlice,
+  argsSlice,
   makeRandom,
 } from './shared';
-import * as CBR from './CBR';
 import {
-  CBR_Null,
-  CBR_Bool,
-  CBR_UInt,
-  CBR_Bytes,
-  CBR_Address,
-  CBR_Digest,
-  CBR_Object,
-  CBR_Data,
-  CBR_Array,
-  CBR_Tuple,
-  CBR_Val,
+  CBR_Address, CBR_Val,
 } from './CBR';
 import waitPort from 'wait-port';
-import { labelMaps, replaceableThunk } from './shared_impl';
+import { replaceableThunk } from './shared_impl';
+import { stdlib as compiledStdlib, ALGO_Ty, NV, typeDefs } from './ALGO_compiled';
 export * from './shared';
 
+
 type BigNumber = ethers.BigNumber;
-const BigNumber = ethers.BigNumber;
-export const UInt_max: BigNumber =
-  BigNumber.from(2).pow(64).sub(1);
+
+type AnyALGO_Ty = ALGO_Ty<CBR_Val>;
+
+export const { addressEq, digest } = compiledStdlib;
+
+export const { T_Null, T_Bool, T_UInt, T_Tuple, T_Array, T_Object, T_Data, T_Bytes, T_Address, T_Digest } = typeDefs;
+
 export const { randomUInt, hasRandom } = makeRandom(8);
 
 // Note: if you want your programs to exit fail
@@ -139,6 +138,7 @@ type ContractAttached = {
   wait: (...argz: any) => any,
   iam: (some_addr: any) => any,
   selfAddress: () => CBR_Address, // Not RawAddress!
+  stdlib: Object,
 };
 
 // TODO
@@ -164,207 +164,6 @@ type ContractInfo = {
 //   recv: function
 // }
 
-// NV = Net Value
-type NV = Uint8Array;
-type ALGO_Ty<BV extends CBR_Val> = {
-  name: string,
-  canonicalize: (uv: unknown) => BV,
-  netSize: number
-  toNet(bv: BV): NV,
-  fromNet(nv: NV): BV,
-}
-type AnyALGO_Ty = ALGO_Ty<CBR_Val>;
-
-export const digest = makeDigest((t:ALGO_Ty<any>, v:any) => t.toNet(v));
-
-export const T_Null: ALGO_Ty<CBR_Null> = {
-  ...CBR.BT_Null,
-  netSize: 0,
-  toNet: (bv: CBR_Null): NV => (void(bv), new Uint8Array([])),
-  fromNet: (nv: NV): CBR_Null => (void(nv), null),
-}
-
-export const T_Bool: ALGO_Ty<CBR_Bool> = {
-  ...CBR.BT_Bool,
-  netSize: 1,
-  toNet: (bv: CBR_Bool): NV => new Uint8Array([bv ? 1 : 0]),
-  fromNet: (nv: NV): CBR_Bool => nv[0] == 1,
-}
-
-export const T_UInt: ALGO_Ty<CBR_UInt> = {
-  ...CBR.BT_UInt,
-  netSize: 8, // UInt64
-  toNet: (bv: CBR_UInt): NV => (
-    ethers.utils.zeroPad(ethers.utils.arrayify(bv), 8)
-  ),
-  fromNet: (nv: NV): CBR_UInt => {
-    // debug(`fromNet: UInt`);
-    // if (getDEBUG()) console.log(nv);
-    return ethers.BigNumber.from(nv);
-  },
-}
-
-/** @description For arbitrary utf8 strings */
-const stringyNet = {
-  toNet: (bv: CBR_Bytes): NV => (
-    ethers.utils.toUtf8Bytes(bv)
-  ),
-  fromNet: (nv: NV): CBR_Bytes => (
-    ethers.utils.toUtf8String(nv)
-  ),
-}
-
-/** @description For hex strings representing bytes */
-const bytestringyNet = {
-  toNet: (bv: string): NV => (
-    ethers.utils.arrayify(bv)
-  ),
-  fromNet: (nv: NV): string => (
-    ethers.utils.hexlify(nv)
-  )
-};
-
-export const T_Bytes = (len:number): ALGO_Ty<CBR_Bytes> => ({
-  ...CBR.BT_Bytes(len),
-  ...stringyNet,
-  netSize: len,
-});
-
-export const T_Digest: ALGO_Ty<CBR_Digest> = {
-  ...CBR.BT_Digest,
-  ...bytestringyNet,
-  netSize: 32,
-};
-
-function addressUnwrapper(x: any): string {
-  const addr =
-    x && x.networkAccount && x.networkAccount.addr
-    || x && x.addr;
-
-  return (addr != undefined)
-    ? '0x' + Buffer.from(algosdk.decodeAddress(addr).publicKey).toString('hex')
-    : x;
-}
-export const T_Address: ALGO_Ty<CBR_Address> = {
-  ...CBR.BT_Address,
-  ...bytestringyNet,
-  netSize: 32,
-  canonicalize: (uv: unknown): CBR_Address => {
-    const val = addressUnwrapper(uv);
-    return CBR.BT_Address.canonicalize(val || uv)
-  }
-}
-
-export const T_Array = (
-  co: ALGO_Ty<CBR_Val>,
-  size: number,
-): ALGO_Ty<CBR_Array> => ({
-  ...CBR.BT_Array(co, size),
-  netSize: size * co.netSize,
-  toNet: (bv: CBR_Array): NV => {
-    return ethers.utils.concat(bv.map((v) => co.toNet(v)));
-  },
-  fromNet: (nv: NV): CBR_Array => {
-    const chunks = new Array(size).fill(null);
-    let rest = nv;
-    for (const i in chunks) {
-      chunks[i] = co.fromNet(rest.slice(0, co.netSize));
-      rest = rest.slice(co.netSize);
-    }
-    // TODO: assert size of nv/rest is correct?
-    return chunks;
-  },
-});
-
-export const T_Tuple = (
-  cos: Array<ALGO_Ty<CBR_Val>>,
-): ALGO_Ty<CBR_Tuple> => ({
-  ...CBR.BT_Tuple(cos),
-  netSize: (
-    cos.reduce((acc, co) => acc + co.netSize, 0)
-  ),
-  toNet: (bv: CBR_Tuple): NV => {
-    const val = cos.map((co, i) => co.toNet(bv[i]));
-    return ethers.utils.concat(val);
-  },
-  // TODO: share more code w/ T_Array.fromNet
-  fromNet: (nv: NV): CBR_Tuple => {
-    const chunks: Array<CBR_Val> = new Array(cos.length).fill(null);
-    let rest = nv;
-    for (const i in cos) {
-      const co = cos[i];
-      chunks[i] = co.fromNet(rest.slice(0, co.netSize));
-      rest = rest.slice(co.netSize);
-    }
-    return chunks;
-  },
-});
-
-export const T_Object = (
-  coMap: {[key: string]: ALGO_Ty<CBR_Val>}
-): ALGO_Ty<CBR_Object> => {
-  const cos = Object.values(coMap);
-  const netSize =
-    cos.reduce((acc, co) => acc + co.netSize, 0);
-  const {ascLabels} = labelMaps(coMap);
-  return {
-    ...CBR.BT_Object(coMap),
-    netSize,
-    toNet: (bv: CBR_Object): NV => {
-      const chunks = ascLabels.map((label) =>
-        coMap[label].toNet(bv[label])
-      );
-      return ethers.utils.concat(chunks);
-    },
-    // TODO: share more code w/ T_Array.fromNet and T_Tuple.fromNet
-    fromNet: (nv: NV): CBR_Object => {
-      const obj: {[key: string]: CBR_Val} = {};
-      let rest = nv;
-      for (const iStr in ascLabels) {
-        const i = parseInt(iStr);
-        const label = ascLabels[i];
-        const co = coMap[label];
-        obj[label] = co.fromNet(rest.slice(0, co.netSize));
-        rest = rest.slice(co.netSize);
-      }
-      return obj;
-    },
-  }
-};
-
-// 1 byte for the label
-// the rest right-padded with zeroes
-// up to the size of the largest variant
-export const T_Data = (
-  coMap: {[key: string]: ALGO_Ty<CBR_Val>}
-): ALGO_Ty<CBR_Data> => {
-  const cos = Object.values(coMap);
-  const valSize =
-    Math.max(...cos.map((co) => co.netSize))
-  const netSize = valSize + 1;
-  const {ascLabels, labelMap} = labelMaps(coMap);
-  return {
-    ...CBR.BT_Data(coMap),
-    netSize,
-    toNet: ([label, val]: CBR_Data): NV => {
-      const i = labelMap[label];
-      const lab_nv = new Uint8Array([i]);
-      const val_co = coMap[label];
-      const val_nv = val_co.toNet(val);
-      const padding = new Uint8Array(valSize - val_nv.length);
-      return ethers.utils.concat([lab_nv, val_nv, padding]);
-    },
-    fromNet: (nv: NV): CBR_Data => {
-      const i = nv[0];
-      const label = ascLabels[i];
-      const val_co = coMap[label];
-      const rest = nv.slice(1);
-      const sliceTo = val_co.netSize;
-      const val = val_co.fromNet(rest.slice(0, sliceTo));
-      return [label, val];
-    },
-  }
-}
 
 // Common interface exports
 
@@ -926,7 +725,7 @@ export const connectAccount = async (networkAccount: NetworkAccount) => {
       }
     };
 
-    return { getInfo, sendrecv, recv, iam, selfAddress, wait };
+    return { getInfo, sendrecv, recv, iam, selfAddress, wait, stdlib: compiledStdlib };
   };
 
   const deployP = async (bin: Backend): Promise<ContractAttached> => {
@@ -1030,6 +829,7 @@ export const connectAccount = async (networkAccount: NetworkAccount) => {
       wait: async(...args: any) => (await implP).wait(...args),
       iam, // doesn't need to await the implP
       selfAddress, // doesn't need to await the implP
+      stdlib: compiledStdlib,
     }
   };
 
@@ -1040,7 +840,7 @@ export const connectAccount = async (networkAccount: NetworkAccount) => {
   const deploy = (bin: Backend): ContractAttached => {
     return deferP(deployP(bin));
   };
-  return { deploy, attach, networkAccount };
+  return { deploy, attach, networkAccount, stdlib: compiledStdlib };
 };
 
 export const balanceOf = async (acc: Account): Promise<BigNumber> => {
@@ -1178,4 +978,3 @@ export const verifyContract = async (ctcInfo: ContractInfo, backend: Backend): P
 
   return true;
 }
-export const addressEq = mkAddressEq(T_Address);
