+++ js/tsconfig.json
@@ -66,5 +66,8 @@
     "resolveJsonModule": false,                /* Include modules imported with '.json' extension */
     "skipLibCheck": true,                     /* Skip type checking of declaration files. */
     "forceConsistentCasingInFileNames": true  /* Disallow inconsistently-cased references to the same file. */
-  }
+  },
+  "include": [
+    "./ts/**/*"
+  ]
 }
